requirements.txt
# =====================================================
# Cybersecurity Assessment Tool - Python Dependencies
# =====================================================
# This file is used by Heroku and pip for deployments.
# To install locally: pip install -r requirements.txt

# ----- Django & Extensions -----
django==5.2.6
djangorestframework==3.16.1
django-appconf==1.2.0
django-debug-toolbar==6.1.0
django-fernet-encrypted-fields==0.3.1
django-fernet-fields==0.6
django-recurrence==1.11.1
django-extensions==4.1
django-picklefield==3.4.0
django-q2==1.9.0

# ----- Heroku / Deployment -----
gunicorn>=22.0,<23.0
whitenoise[brotli]>=6.7,<7.0
dj-database-url==3.1.2

# ----- Database -----
psycopg2-binary==2.9.10
psycopg[binary]==3.2.10

# ----- Google / Gemini AI -----
google-genai==1.39.1
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2

# ----- HTTP / Networking -----
httpx==0.28.1
httpcore==1.0.9
httplib2==0.31.0
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# ----- Data / Serialization -----
pydantic==2.11.9
pydantic-core==2.33.2
annotated-types==0.7.0
typing-inspection==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
Markdown==3.9
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0

# ----- PDF & Document Generation -----
pypdf==6.1.3
pypandoc==1.15
pandoc==2.4
pydyf==0.11.0
cssselect2==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
webencodings==0.5.1
pyphen==0.17.2

# ----- Data Science / Plotting -----
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.8
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.5.0

# ----- Testing -----
Faker==37.12.0

# ----- Crypto / Encoding -----
cryptography==46.0.5
cffi==2.0.0
pycparser==2.23
Brotli==1.1.0
zopfli==0.2.3.post1

# ----- Utils -----
python-dotenv==1.1.1
pillow==11.3.0
sqlparse==0.5.3
asgiref==3.9.2
tenacity==9.1.2
tqdm==4.67.1
cachetools==5.5.2
pyparsing==3.2.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
rsa==4.9.1
sniffio==1.3.1
anyio==4.11.0
h11==0.16.0
uritemplate==4.2.0
websockets==15.0.1
tzdata==2025.2
pytz==2025.2
plumbum==1.10.0
ply==3.11
arrow==1.4.0
blessed==1.33.0
jinxed==1.3.0
wcwidth==0.6.0